Creates the HVAC_adaptive service
It adjusts one of my thermostat program to free as much power
bandwidth as possible. This service postpones this particular
thermostat program if it estimates that it will take less time to
reach the target temperature.

The estimation is based on a database built out of the thermal
property of my home derived from the `sensor.csv' files constructed
by the `sensor_logger' service.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>