EV_adaptive_charging: Introduce the `maximize' setting
By default, `maximize' is equal to False. If this setting is set to
True the EV charger maximizes the available power usage. For example:
with an available power of 1.56 KW the EV charger charges at 6A which
mean 1.44KW. However, if maximize is True, it charges at 7A (1.68 KW)
because 1.56 KW is closer to 1.68 KW than to 1.44 KW.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>