water_heater: do not call `is_runnable` from `meet_running_criteria`
`meet_running_criteria` is about deciding if the task could be started
or could keep running based on supplied power consumption data and
some internal to the task conditions. `is_runnable` is about deciding
if it is okay to start the task.

These two functions should not call each other.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>