weather: improve robustness and report forecast data update issue
As soon as the forecast data is getting outdated, the weather module
report the issue to the Monitor. Usually, the weather forecast data
retrieved from the server covers several days. Nevertheless a couple
of module depends on this data and it is critical to be aware of any
issue as soon as possible to implement corrective action this lack of
data contaminates other components.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>