water_heater: improve task efficiency
This patch makes the water_heater task better detects situation when
the water_heater does not need to run. If the water heater has been
running for a little while, meaning it was using power, and it stops
using power the water_heater is made unrunnable for a larger period of
time.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>