sensor_logger: code sharing and robustness
When data is inserted to the database there can be a precision loss in
particular with floats. As the result, when the data is read back from
the database and compared to a new record sensor_logger may wrongly
detect a change in the data and unnecessarily insert a new row.

This patch keeps the latest data in a local dictionary from comparison
purposes instead of pulling the information from the database.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>