car_charger: Handle multiple cars
With the replacement of our latest remaining gas car with a Model 3
2020 LR, we had to make the charger task handle both cars.

The task needs to know which car is connected in order to retrieve the
state of charge and decide of the appropriate priority. Also the usage
of these two cars is very different: the Model 3 is our long distance
car that we use very occasionally during the week and the Bold EV is
our commute car that we use all the time. As result, their priority
policies are different.

The implementation relies on the fact that the Model 3 car sensor is
able to report where the car is located in the garage and if the
charger port is connected. With this information, the car_charger task
can decide whether the Model 3 is connected or the Bold EV.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>