scheduler: do not allow concurrent same power keys tasks
The power consumption of tasks sharing the same keys cannot be clearly
identified. Therefor, they do not run at the same time.

I am not convinced this is the right way to address it but it is a
simple to make sure the various car chargers which share the same plug
do not overload the circuit by running simultaneously.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>