hvac: double estimate time
It turns out it is preferable to over estimate the running time. If
the programmed time was under-estimate, the HVAC system turns off
before reaching the target and has to be turned back on. It creates
some waste of energy. Additionally, it diminishes the chance to reach
the target as the HVAC system won't turn right back on to prevent part
damage. It usually takes a few minutes before HVAC successfully turns
on again. All this time is also wasted.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>