hvac: do not make hvac reckless when its priority reached URGENT
The current implementation is designed to make sure the hvac turns on
to reach its goal by accepting to run regardless of the available
power. As a result, we observe some over consumption at this point of
priority transition as other task may still have to complete their
minimum cycle.

In order to smooth the transition while guaranteeing the hvac will
turn on as soon as possible, the `meet_running_criteria' is
quadratically considering the remaining time over the estimate time.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>